Made entities integration test suite and other minor changes

+++ src/main/java/com/backend/domain/AvailableCarService.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -26,7 +27,7 @@ public class AvailableCarService {
     private String description;
 
     @Column(name = "cost", nullable = false)
-    private Long cost;
+    private BigDecimal cost;
 
     @Column(name = "repair_time")
     private int repairTimeInMinutes;
@@ -35,13 +36,13 @@ public class AvailableCarService {
     private String premiumMakes;
 
     @Column(name = "make_multiplier")
-    private Long makeMultiplier;
+    private BigDecimal makeMultiplier;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "garage_id")
     private Garage garage;
 
-    public AvailableCarService(Long id, String name, String description, Long cost, int repairTimeInMinutes, String premiumMakes, Long makeMultiplier) {
+    public AvailableCarService(Long id, String name, String description, BigDecimal cost, int repairTimeInMinutes, String premiumMakes, BigDecimal makeMultiplier) {
         this.id = id;
         this.name = name;
         this.description = description;
