+++ src/main/java/com/backend/domain/dto/AvailableCarServiceDto.java
@@ -3,6 +3,7 @@ package com.backend.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Time;
 
 @Data
@@ -11,8 +12,8 @@ public class AvailableCarServiceDto {
     private Long id;
     private String name;
     private String description;
-    private Long cost;
+    private BigDecimal cost;
     private int repairTimeInMinutes;
     private String premiumMakes;
-    private Long makeMultiplier;
+    private BigDecimal makeMultiplier;
 }
